+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -209,6 +209,7 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
   }
 
   public async leaveCommunity() {
+    this.tor.resetHiddenServices()
     this.serverIoProvider.io.close()
     await this.localDbService.purge()
     await this.closeAllServices({ saveTor: true })