+++ packages/backend/src/nest/libp2p/libp2p.utils.ts
@@ -1,7 +1,11 @@
+const ONION = '.onion'
+
 export const createLibp2pAddress = (address: string, peerId: string) => {
-  return `/dns4/${address}/tcp/80/ws/p2p/${peerId}`
+  if (!address.endsWith(ONION)) address += ONION
+  return `/dns4/${address}/tcp/80/wss/p2p/${peerId}`
 }
 
 export const createLibp2pListenAddress = (address: string) => {
-  return `/dns4/${address}/tcp/80/ws`
+  if (!address.endsWith(ONION)) address += ONION
+  return `/dns4/${address}/tcp/80/wss`
 }