+++ packages/desktop/src/renderer/components/ChannelCreationModal/ChannelCreationModal.test.tsx
@@ -13,8 +13,9 @@ describe('Create ChannelCreationModalComponent', () => {
           aria-hidden="true"
         />
         <div
-          class="Modalroot MuiModal-root css-1voaj9u-MuiModal-root"
+          class="MuiModal-root css-1vjugmr-MuiModal-root"
           role="presentation"
+          zindex="1300"
         >
           <div
             aria-hidden="true"