+++ packages/backend-bundle/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.6.0](/compare/backend-bundle@1.6.0-alpha.0...backend-bundle@1.6.0) (2023-08-28)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
+# [1.6.0-alpha.0](https://github.com/TryQuiet/quiet/compare/backend-bundle@1.5.0...backend-bundle@1.6.0-alpha.0) (2023-08-25)
+
+**Note:** Version bump only for package backend-bundle
+
+
+
+
+
 # [1.5.0](https://github.com/TryQuiet/quiet/compare/backend-bundle@1.4.0...backend-bundle@1.5.0) (2023-08-17)
 
 **Note:** Version bump only for package backend-bundle